// ABC/CC/UI/src/types/GetSiteDetailsRequestModel.ts
export interface GetSiteDetailsRequestModel {
    providerCity: string | null;
    providerEligibilityType: string | null;
    providerNPI: string | null;
    providerName: string | null;
    providerOAOID: string | null;
    providerTIN: string | null;
    providerZip: string | null;
    requestId: string | null;
    memberIPA: string| null;
}